admin-module/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis
redis[hiredis]==5.2.1
redis-sentinel==1.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0
python-ldap==3.4.4
python-decouple==3.8

# Validation
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Monitoring
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-prometheus==0.50b0
opentelemetry-exporter-jaeger==1.27.0

# Configuration
pyyaml==6.0.2

# Logging
python-json-logger==3.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1

# Development
black==24.10.0
flake8==7.1.1
mypy==1.14.1
isort==5.13.2

# Distributed Consensus (Raft)
raft-python==0.3.0

# Vector Clocks
vector-clock==0.1.0

# Kafka for async processing
aiokafka==0.12.0
